Refactor util from single file to modular setup
