+++ src/utils/doc.h
@@ -1,0 +1,9 @@
+#pragma once
+
+// This file only exists for the doxygen namespace comment.
+
+/**
+ * @namespace util
+ * @brief A collection of utility functions and classes with common use-cases.
+ */
+namespace util {};
